Fixed display bug in Warnins and Errors window
