+++ src/main/java/com/cburch/logisim/fpga/fpgagui/FPGACommanderListModel.java
@@ -30,140 +30,34 @@
 
 package com.cburch.logisim.fpga.fpgagui;
 
-import static com.cburch.logisim.fpga.Strings.S;
-
-import java.awt.Color;
-import java.awt.Component;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Set;
 
 import javax.swing.AbstractListModel;
-import javax.swing.JLabel;
-import javax.swing.JList;
 import javax.swing.ListCellRenderer;
 import javax.swing.event.ListDataEvent;
 import javax.swing.event.ListDataListener;
 
 import com.cburch.logisim.fpga.designrulecheck.SimpleDRCContainer;
-import com.cburch.logisim.util.Icons;
+import com.cburch.logisim.fpga.gui.ListModelCellRenderer;
 
 @SuppressWarnings("serial")
 public class FPGACommanderListModel extends  AbstractListModel<Object> {
 	
-	private static Color FATAL = Color.RED;
-	private static Color SEVERE = Color.yellow;
-	private static Color NORMAL = Color.LIGHT_GRAY;
-	private static Color ADDENDUM = Color.GRAY;
-
-	public class ListModelCellRenderer extends JLabel implements ListCellRenderer<Object> {
-		
-		private boolean CountLines;
-		
-        
-		public ListModelCellRenderer(boolean countLines) {
-			CountLines = countLines;
-		}
-		
-		@Override
-		public Component getListCellRendererComponent(
-				JList<? extends Object> list, 
-				Object value, 
-				int index,
-				boolean isSelected, 
-				boolean cellHasFocus) {
-			SimpleDRCContainer msg = null;
-			setBackground(list.getBackground());
-			setForeground(list.getForeground());
-			StringBuffer Line = new StringBuffer();
-			setIcon(Icons.getIcon("empty.png")); /* place holder too make space for the trace icon */
-			if (value instanceof SimpleDRCContainer) {
-				msg = (SimpleDRCContainer) value;
-			}
-			if (msg != null) {
-				if (msg.DRCInfoPresent()) {
-		        	setIcon(Icons.getIcon("drc_trace.png"));
-				} 
-				switch (msg.Severity()) {
-					case SimpleDRCContainer.LEVEL_SEVERE :
-						setForeground(SEVERE);
-						break;
-					case SimpleDRCContainer.LEVEL_FATAL :
-						setBackground(FATAL);
-						setForeground(list.getBackground());
-						break;
-					default : 
-						setForeground(NORMAL);
-				}
-			}
-			if (value.toString().contains("BUG")) {
-				setBackground(Color.MAGENTA);
-				setForeground(Color.black);
-			}
-			if (CountLines) {
-				if (msg != null) {
-					if (msg.SupressCount()) {
-						setForeground(ADDENDUM);
-						Line.append("       ");
-					} else {
-						int line = msg.GetListNumber();
-						if (line < 10) {
-							Line.append("    ");
-						} else if (line < 100) {
-							Line.append("   ");
-						} else if (line < 1000) {
-							Line.append("  ");
-						} else if (line < 10000) {
-							Line.append(" ");
-						}
-						Line.append(Integer.toString(line) + "> ");
-					}
-				} else {
-					if (index < 9) {
-						Line.append("    ");
-					} else if (index < 99) {
-						Line.append("   ");
-					} else if (index < 999) {
-						Line.append("  ");
-					} else if (index < 9999) {
-						Line.append(" ");
-					}
-					Line.append(Integer.toString(index + 1) + "> ");
-				}
-			}
-			if (msg != null) {
-				switch (msg.Severity()) {
-					case SimpleDRCContainer.LEVEL_SEVERE :
-						Line.append(S.get("SEVERE_MSG")+" ");
-						break;
-					case SimpleDRCContainer.LEVEL_FATAL :
-						Line.append(S.get("FATAL_MSG")+" ");
-						break;
-				}
-				if (msg.HasCircuit()) {
-					Line.append(msg.GetCircuit().getName()+": ");
-				}
-			}
-			Line.append(value.toString());
-			setText(Line.toString());
-			setEnabled(list.isEnabled());
-			setFont(list.getFont());
-			setOpaque(true);
-			return this;
-		}
-	}
-
 	private ArrayList<Object> myData;
 	private Set<ListDataListener> myListeners;
 	private int count = 0;
+	private ListModelCellRenderer MyRender;
 	
-	public FPGACommanderListModel() {
+	public FPGACommanderListModel(boolean CountLines) {
 		myData = new ArrayList<Object>();
 		myListeners = new HashSet<ListDataListener>();
+		MyRender = new ListModelCellRenderer(CountLines);
 	}
 	
-	public ListCellRenderer<Object> getMyRenderer(boolean CountLines) {
-		return new ListModelCellRenderer(CountLines);
+	public ListCellRenderer<Object> getMyRenderer() {
+		return MyRender;
 	}
 	
 	public void clear() {
